feat(UserController, UserService, UserServiceImpl, SecurityConfig, JwtAuthFilter, AuthController): implement getAllUsers endpoint restricted to veterinarians with role-based validation
